+++ src/compas_cra/equilibrium/cra_penalty_pyomo.py
@@ -6,6 +6,7 @@ import numpy as np
 import pyomo.environ as pyo
 from compas_assembly.datastructures import Assembly
 
+from ._solver import ipopt_solver
 from .cra_helper import equilibrium_setup
 from .cra_helper import external_force_setup
 from .cra_helper import friction_setup
@@ -137,12 +138,13 @@ def cra_penalty_solve(
     if timer:
         start_time = time.time()
 
-    solver = pyo.SolverFactory("ipopt")
-    solver.options["tol"] = 1e-8  # same as default tolerance
-    solver.options["constr_viol_tol"] = 1e-7  # constraint tolerance
-    solver.options["acceptable_tol"] = 1e-6
-    solver.options["acceptable_constr_viol_tol"] = 1e-5
     # https://coin-or.github.io/Ipopt/OPTIONS.html
+    solver = ipopt_solver(
+        tol=1e-8,  # same as default tolerance
+        constr_viol_tol=1e-7,  # constraint tolerance
+        acceptable_tol=1e-6,
+        acceptable_constr_viol_tol=1e-5,
+    )
     result = solver.solve(model, tee=verbose)
 
     if timer: